Ensure full CSS file is included in hash
