+++ src/css/styles.11ty.js
@@ -1,33 +1,15 @@
-const fs = require('fs');
-const path = require('path');
-const postcss = require('postcss');
-const cssFilename = require('../utils/css-filename');
-
-// the file name as an entry point for postcss compilation
-// also used to define the output filename in our output /css folder.
-const fileName = 'app.css';
+const manageStyles = require('../utils/manage-styles');
 
 module.exports = class {
-  data() {
-    const rawFilepath = path.join(__dirname, `../_includes/css/${fileName}`);
-    const rawCss = fs.readFileSync(rawFilepath);
-    const filename = cssFilename();
+  async data() {
+    const permalink = await manageStyles.getFilename();
 
     return {
-      permalink: filename,
-      rawFilepath,
-      rawCss,
+      permalink,
     };
   }
 
-  async render({ rawCss, rawFilepath }) {
-    return await postcss([
-      // require('postcss-comment'),
-      require('precss'),
-      require('postcss-import'),
-      require('cssnano'),
-    ])
-      .process(rawCss, { from: rawFilepath })
-      .then((result) => result.css);
+  async render() {
+    return await manageStyles.getContents();
   }
 };
